requirements.txt
fastapi
appwrite
pydantic
python-dotenv
twilio
google.generativeai
google-adk
python-multipart
uvicorn
duckduckgo-search